requirements.txt
aniso8601==7.0.0
backports.csv==1.0.7
beautifulsoup4==4.8.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-bootstrap4==1.0.1
django-graphiql==0.4.4
django-import-export==1.2.0
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
jdcal==1.4.1
odfpy==1.4.0
openpyxl==3.0.0
pkg-resources==0.0.0
promise==2.2.1
pytz==2019.2
PyYAML==5.1.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
